Add support for multiple controllers
