+++ fw/bpm.h
@@ -7,8 +7,7 @@ class Bpm {
 public:
   explicit Bpm(uint8_t max_level);
 
-  uint8_t get() const;
-  void update(uint16_t button_state);
+  uint8_t update(uint16_t button_state);
 
 private:
   uint8_t max_level;
