+++ fw/devices/iidx/iidx_combo.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../combo.h"
+
+namespace Iidx {
+  combo get_button_combo(uint16_t button_state);
+}
