+++ fw/devices/sdvx/sdvx_combo.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../combo.h"
+
+namespace Sdvx {
+  combo get_button_combo(uint16_t button_state);
+}
