+++ fw/devices/sdvx/sdvx_rgb.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stdint.h>
+
+namespace Sdvx {
+  struct hid_lights {
+    uint16_t buttons;
+  };
+}
