+++ fw/devices/sdvx/sdvx_rgb_manager.cpp
@@ -1,0 +1,6 @@
+#include "sdvx_rgb_manager.h"
+
+namespace Sdvx {
+  namespace RgbManager {
+  }
+}
