+++ fw/devices/sdvx/sdvx_rgb_manager.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace Sdvx {
+  namespace RgbManager {
+  }
+}
