+++ fw/devices/sdvx/sdvx_usb.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../config.h"
+
+namespace Sdvx {
+  void usb_init(config &config);
+}
