+++ fw/pin.h
@@ -28,6 +28,11 @@ typedef struct {
   uint16_t tt_position;
 } tt_pins;
 
+typedef struct {
+  uint8_t pin;
+  uint8_t position;
+} encoder_pin;
+
 extern hw_pin hw_pins[6];
 
 // to index into hw_pins[] array
